pyproject.toml
[tool.poetry]
name = "arkn98"
version = "0.1.0"
description = ""
authors = ["arkn98 <20590666+arkn98@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
